Vendor real jQueryInitialize (pre/post 3.2.6) and run it in the harness
Restructure trigger-performance from a static demo into a source/ build
project that bundles the REAL jQueryInitialize from pixiebrix-source — both
the pre-3.2.6 and 3.2.6 implementations — so the pages exercise the actual
code path a customer reported crashing, not a simulation.

- source/trigger-performance/vendor/{new,old}/: verbatim source files
  (jQueryInitialize + sharedMutationObserver + selectorTokenIndex +
  routeChange for new; jQueryInitialize for old) with a minimal $safeFind
  shim. Relative imports de-extensioned for esbuild; jQuery pinned to 3.7.1
  for Sizzle's $.find.tokenize.
- build.sh + bundle.mjs (esbuild) emit window.PBXVendor.{legacy,optimized}.
- harness.js now drives the real initialize() and reports black-box crash
  signals: FPS, long-tasks/sec, max task time, callback matches.
- Pages moved to source/.../pages and emitted to public/trigger-performance.

Verified in headless Chrome: real old-vs-new contrast is deterministic,
e.g. detached-subtree legacy = 28k matches (scans detached portals) vs
optimized = 0 (the isConnected fast-path from #8170).
