+++ public/trigger-performance/attribute-cascade.html
@@ -15,6 +15,8 @@
   </style>
 </head>
 <body>
+  <script src="vendor/jqi-old.js"></script>
+  <script src="vendor/jqi-new.js"></script>
   <script src="harness.js"></script>
   <script>
     PBX.page({
