+++ public/trigger-performance/deep-dom.html
@@ -14,6 +14,8 @@
   </style>
 </head>
 <body>
+  <script src="vendor/jqi-old.js"></script>
+  <script src="vendor/jqi-new.js"></script>
   <script src="harness.js"></script>
   <script>
     PBX.page({
