+++ public/trigger-performance/detached-subtree.html
@@ -12,6 +12,8 @@
   </style>
 </head>
 <body>
+  <script src="vendor/jqi-old.js"></script>
+  <script src="vendor/jqi-new.js"></script>
   <script src="harness.js"></script>
   <script>
     PBX.page({
