+++ public/trigger-performance/infinite-scroll-burst.html
@@ -13,6 +13,8 @@
   </style>
 </head>
 <body>
+  <script src="vendor/jqi-old.js"></script>
+  <script src="vendor/jqi-new.js"></script>
   <script src="harness.js"></script>
   <script>
     PBX.page({
