+++ public/trigger-performance/many-selectors.html
@@ -12,6 +12,8 @@
   </style>
 </head>
 <body>
+  <script src="vendor/jqi-old.js"></script>
+  <script src="vendor/jqi-new.js"></script>
   <script src="harness.js"></script>
   <script>
     // The number of distinct selectors is itself a tunable, so read it up front
