+++ public/trigger-performance/route-swap.html
@@ -17,6 +17,8 @@
   </style>
 </head>
 <body>
+  <script src="vendor/jqi-old.js"></script>
+  <script src="vendor/jqi-new.js"></script>
   <script src="harness.js"></script>
   <script>
     PBX.page({
