+++ source/trigger-performance/pages/route-swap.html
@@ -1,0 +1,83 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>trigger-performance: route swap</title>
+  <style>
+    body { font-family: system-ui, sans-serif; margin: 0; }
+    #pbx-root { padding: 16px 20px; }
+    button.pbx-act { padding: 6px 12px; margin: 0 6px 12px 0; cursor: pointer;
+      border: 1px solid #2563eb; background: #2563eb; color: #fff; border-radius: 4px; font: inherit; }
+    #route-status { font: 12px monospace; color: #475569; margin-bottom: 10px; }
+    .pii { display: inline-block; width: 90px; height: 16px; margin: 1px;
+      background: #cbd5e1; border-radius: 3px; }
+    body.route-active .pii { background: #fca5a5; } /* ancestor-gated mask target */
+    .filler { display: inline-block; width: 90px; height: 16px; margin: 1px; background: #f1f5f9; }
+  </style>
+</head>
+<body>
+  <script src="vendor/jqi-old.js"></script>
+  <script src="vendor/jqi-new.js"></script>
+  <script src="harness.js"></script>
+  <script>
+    PBX.page({
+      title: "Large SPA route swap",
+      intro: `
+        <h1>Large SPA route swap</h1>
+        <p>Each "route change" calls <code>history.pushState</code> and replaces the
+        whole view with a big batch of sibling nodes in one mutation. Many of them
+        are <code>.pii</code> elements gated on an ancestor class
+        (<code>body.route-active .pii</code>) — the masking pattern PixieBrix
+        triggers use. Exercises the <strong>burst-size flush (PR #8170)</strong> and
+        the <strong>route-change re-sweep (PR #8202, Tier 1S)</strong>.</p>
+        <p><strong>Try it:</strong> keep <code>siblings</code> &gt; 512 to trip the
+        burst threshold. In <code>legacy</code> mode rapid swaps stack up full
+        document scans and stutter; <code>optimized</code> defers the big batch to
+        idle and re-sweeps ancestor-gated targets within one frame of the URL change.</p>
+        <p>Target selector for a real mod: <code>body.route-active .pii</code></p>`,
+      selectors: ["body.route-active .pii"],
+      tunables: [
+        { key: "siblings", label: "Siblings per route", def: 800, min: 10, max: 20000 },
+        { key: "interval", label: "Auto-swap every (ms, 0=off)", def: 0, min: 0, max: 10000 },
+      ],
+      build(root, p) {
+        const bar = document.createElement("div");
+        const btn = document.createElement("button");
+        btn.className = "pbx-act";
+        btn.textContent = "▸ Swap route now";
+        const status = document.createElement("div");
+        status.id = "route-status";
+        const view = document.createElement("div");
+        view.id = "route-view";
+        bar.appendChild(btn);
+        root.append(bar, status, view);
+
+        let route = 0;
+        function swap() {
+          route++;
+          history.pushState({ route }, "", "?route=" + route + location.hash);
+          document.body.classList.add("route-active");
+          // Replace the entire view in a single mutation batch.
+          const frag = document.createDocumentFragment();
+          for (let i = 0; i < p.siblings; i++) {
+            const d = document.createElement("div");
+            d.className = i % 4 === 0 ? "pii" : "filler";
+            frag.appendChild(d);
+          }
+          view.replaceChildren(frag);
+          status.textContent = `route ${route} — swapped ${p.siblings} nodes (${Math.ceil(
+            p.siblings / 4
+          )} .pii)`;
+        }
+        btn.onclick = swap;
+        root._swap = swap;
+        swap();
+      },
+      start(root, p) {
+        if (p.interval > 0) PBX.every(p.interval, () => root._swap());
+      },
+    });
+  </script>
+</body>
+</html>
